Update control.lua
I was getting an error about stack doesn't have name property or something like that on line 180. I'm not a lua programmer but after this i wasn't getting the error any more at least for the rest of the night. Take a look and see what might be causing it.